requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
httpx>=0.27.0
aiofiles>=23.0.0
python-multipart>=0.0.9
jinja2>=3.1.4
pydantic-settings>=2.0.0
